refactor(PowerDisplay): extract DisplayConfigInventory from DdcCiNative
GetAllMonitorDisplayInfo is the Phase 0 entry point used by both DDC/CI
and WMI discovery paths, plus the MainViewModel monitor-number mapping.
Hosting it on DdcCiNative was a historical accident — the function does
not call DDC/CI, only QueryDisplayConfig + DisplayConfigGetDeviceInfo.

Move it to a new DisplayConfigInventory class in the parent Drivers
namespace so its role as a neutral, controller-agnostic enumerator is
explicit, and DdcCiNative is left with only DDC/CI-specific helpers.
