+++ src/modules/powerdisplay/PowerDisplay.Lib/Drivers/DDC/DdcCiNative.cs
@@ -3,11 +3,8 @@
 // See the LICENSE file in the project root for more information.
 
 using System;
-using System.Collections.Generic;
 using System.Runtime.InteropServices;
 using ManagedCommon;
-using Windows.Win32.Foundation;
-using static PowerDisplay.Common.Drivers.NativeConstants;
 using static PowerDisplay.Common.Drivers.PInvoke;
 
 namespace PowerDisplay.Common.Drivers.DDC
@@ -124,154 +121,6 @@ namespace PowerDisplay.Common.Drivers.DDC
                 Logger.LogError($"DDC: TryGetCapabilitiesString exception (handle=0x{hPhysicalMonitor:X}): {ex.Message}");
                 return null;
             }
-        }
-
-        /// <summary>
-        /// Gets GDI device name for a source (e.g., "\\.\DISPLAY1").
-        /// </summary>
-        /// <param name="adapterId">Adapter ID</param>
-        /// <param name="sourceId">Source ID</param>
-        /// <returns>GDI device name, or null if retrieval fails</returns>
-        private static unsafe string? GetSourceGdiDeviceName(LUID adapterId, uint sourceId)
-        {
-            try
-            {
-                var sourceName = new DisplayConfigSourceDeviceName
-                {
-                    Header = new DisplayConfigDeviceInfoHeader
-                    {
-                        Type = DisplayconfigDeviceInfoGetSourceName,
-                        Size = (uint)sizeof(DisplayConfigSourceDeviceName),
-                        AdapterId = adapterId,
-                        Id = sourceId,
-                    },
-                };
-
-                var result = DisplayConfigGetDeviceInfo(&sourceName);
-                if (result == 0)
-                {
-                    return sourceName.GetViewGdiDeviceName();
-                }
-            }
-            catch (Exception ex) when (ex is not OutOfMemoryException)
-            {
-            }
-
-            return null;
-        }
-
-        /// <summary>
-        /// Gets friendly name, device path, and output technology for a monitor target.
-        /// </summary>
-        /// <param name="adapterId">Adapter ID</param>
-        /// <param name="targetId">Target ID</param>
-        /// <returns>Tuple of (friendlyName, devicePath, outputTechnology), values may be null/0 if retrieval fails</returns>
-        private static unsafe (string? FriendlyName, string? DevicePath, uint OutputTechnology) GetTargetDeviceInfo(LUID adapterId, uint targetId)
-        {
-            try
-            {
-                var deviceName = new DisplayConfigTargetDeviceName
-                {
-                    Header = new DisplayConfigDeviceInfoHeader
-                    {
-                        Type = DisplayconfigDeviceInfoGetTargetName,
-                        Size = (uint)sizeof(DisplayConfigTargetDeviceName),
-                        AdapterId = adapterId,
-                        Id = targetId,
-                    },
-                };
-
-                var result = DisplayConfigGetDeviceInfo(&deviceName);
-                if (result == 0)
-                {
-                    return (
-                        deviceName.GetMonitorFriendlyDeviceName(),
-                        deviceName.GetMonitorDevicePath(),
-                        deviceName.OutputTechnology);
-                }
-            }
-            catch (Exception ex) when (ex is not OutOfMemoryException)
-            {
-            }
-
-            return (null, null, 0u);
-        }
-
-        /// <summary>
-        /// Gets complete information for all monitors, keyed by GDI device name (e.g., "\\.\DISPLAY1").
-        /// This allows reliable matching with GetMonitorInfo results.
-        /// </summary>
-        /// <returns>Dictionary keyed by GDI device name containing monitor information</returns>
-        public static unsafe Dictionary<string, MonitorDisplayInfo> GetAllMonitorDisplayInfo()
-        {
-            var monitorInfo = new Dictionary<string, MonitorDisplayInfo>(StringComparer.OrdinalIgnoreCase);
-
-            try
-            {
-                // Get buffer sizes
-                var result = GetDisplayConfigBufferSizes(QdcOnlyActivePaths, out uint pathCount, out uint modeCount);
-                if (result != 0)
-                {
-                    return monitorInfo;
-                }
-
-                // Allocate buffers
-                var paths = new DisplayConfigPathInfo[pathCount];
-                var modes = new DisplayConfigModeInfo[modeCount];
-
-                // Query display configuration using fixed pointer
-                fixed (DisplayConfigPathInfo* pathsPtr = paths)
-                {
-                    fixed (DisplayConfigModeInfo* modesPtr = modes)
-                    {
-                        result = QueryDisplayConfig(QdcOnlyActivePaths, ref pathCount, pathsPtr, ref modeCount, modesPtr, IntPtr.Zero);
-                        if (result != 0)
-                        {
-                            return monitorInfo;
-                        }
-                    }
-                }
-
-                // Get information for each path
-                // The path index corresponds to Windows Display Settings "Identify" number
-                for (int i = 0; i < pathCount; i++)
-                {
-                    var path = paths[i];
-
-                    // Get GDI device name from source info (e.g., "\\.\DISPLAY1")
-                    var gdiDeviceName = GetSourceGdiDeviceName(path.SourceInfo.AdapterId, path.SourceInfo.Id);
-                    if (string.IsNullOrEmpty(gdiDeviceName))
-                    {
-                        continue;
-                    }
-
-                    // Get target info (friendly name, device path, output technology)
-                    var (friendlyName, devicePath, outputTechnology) = GetTargetDeviceInfo(path.TargetInfo.AdapterId, path.TargetInfo.Id);
-
-                    // Use device path as key - unique per target, supports mirror mode
-                    if (string.IsNullOrEmpty(devicePath))
-                    {
-                        continue;
-                    }
-
-                    monitorInfo[devicePath] = new MonitorDisplayInfo
-                    {
-                        DevicePath = devicePath,
-                        GdiDeviceName = gdiDeviceName,
-                        FriendlyName = friendlyName ?? string.Empty,
-                        AdapterId = path.TargetInfo.AdapterId,
-                        TargetId = path.TargetInfo.Id,
-                        MonitorNumber = i + 1, // 1-based, matches Windows Display Settings
-                        OutputTechnology = outputTechnology,
-                        IsInternal = DisplayClassifier.IsInternal(outputTechnology),
-                    };
-                }
-            }
-            catch (Exception ex) when (ex is not OutOfMemoryException)
-            {
-            }
-
-            return monitorInfo;
         }
     }
 }
